Изменено поле id у User, соответсвенно, поменены параметры у методов юзера. Изменено имя колонки с fk_user_id на user_id, соответственно, везде поменено в классах Post это замечание. В User добавлено поле @OneToMany List<Post> posts для хранения созданных пользователем постов. Импорты со звёздочкой заменены на стандартные, без подтяга.

+++ src/main/java/com/corporation/service/UserService.java
@@ -18,7 +18,7 @@ public class UserService {
     private final UserRepository userRepository;
 
     @Transactional
-    public Optional<User> findById(int id) {
+    public Optional<User> findById(long id) {
         return userRepository.findById(id);
     }
 }
